+++ packages/editor-core/test/browser/ui/ChromeExpanded.tsx
@@ -55,6 +55,24 @@ describe('@atlaskit/editor-core/ui/ChromeExpanded', () => {
       expect(props['style']!.maxHeight).to.equal('75px');
     });
 
+    it('should not set height if props.height is undefined', () => {
+      const { editorView } = editor(doc(p()));
+      const chrome = mount(<ChromeExpanded editorView={editorView} onSave={noop} saveDisabled={false} />);
+
+      const wrapper = chrome.find(Content).find('div').at(1);
+      const props = wrapper.props()!;
+      expect(props.style!.height).to.equal(undefined);
+    });
+
+    it('should not set borders if only props.height is set', () => {
+      const { editorView } = editor(doc(p()));
+      const chrome = mount(<ChromeExpanded editorView={editorView} onSave={noop} saveDisabled={false} height={75} />);
+
+      const wrapper = chrome.find(Content).find('div').at(1);
+      const props = wrapper.props()!;
+      expect(props.style!.borderBottom).to.equal(undefined);
+    });
+
     it('should render disabled save button if saveDisabled=true', () => {
       const { editorView } = editor(doc(p()));
       const chrome = mount(<ChromeExpanded editorView={editorView} onSave={noop} saveDisabled={true} />);